separate python from singer
