+++ airbyte-integrations/base-python/.dockerignore
@@ -1,0 +1,6 @@
+*
+!Dockerfile
+!base.py
+!destination.py
+!integration.py
+!source.py
