+++ airbyte-integrations/base-python/destination.py
@@ -1,0 +1,6 @@
+from integration import Integration
+
+
+class SingerSource(Integration):
+    def __init__(self):
+        pass
