--- airbyte-integrations/singer/base-singer/singer_destination.py
@@ -1,6 +1,0 @@
-from singer_integration import SingerIntegration
-
-
-class SingerSource(SingerIntegration):
-    def __init__(self):
-        pass
