+++ airbyte-integrations/singer/exchangeratesapi_io/source/Dockerfile
@@ -12,6 +12,6 @@ RUN python -m pip install --upgrade pip && \
   pip install -r requirements.txt
 
 COPY entrypoint.sh .
-COPY source_implementation.py /base_singer
+COPY source_implementation.py /airbyte
 
 LABEL io.airbyte.version=0.1.2
